Move dashing into its own layer
An additional style for the fill layers is not going to work reliably.
Changes to order of the .mss files which seem to have to same meaning
from a style author point of view lead to different results in XML and
make authors become crazy. Splitting of into a separate layer is a bit
more difficult because we need compositing operations to achieve the
same effect but it makes working on the style easier.