Define features to use in popups (#152)
Part of https://github.com/hiddewie/OpenRailwayMap-vector/issues/107

Part of https://github.com/hiddewie/OpenRailwayMap-vector/issues/49

This PR defines features for every source, with a identifier and
description.

Changes:
- Simplify existing feature YAML content (remove one layer of indent).
- Let the proxy host a `features.json` that contains the descriptions
for the feature in each layer.
- Define a few new feature files in the `features` directory that
describe features of a certain layer.
- Use some of the feature files in the legend generation. Some legends
are still layer specific and custom.
- Generate the popup from the structured feature descriptions: title,
properties and type (point, line or polygon) for OSM link generation.
- Add many TODOs in feature code where more properties could be made
available for usage in the popup.

Dynamic icons (e.g. based on signal type or speed values) for signals do
not work yet and will show `undefined` in the label.

*View diff with whitespace disabled*