Unknown signals (#203)
Fixes #141 

This pull request adds support for rendering railway signals annotated
with one of the categories of signals, but not mapped to a known
feature.

The icons are pentagons with a border of a certain color, to make them
stand out compared to the real signal icons.

Unknown speed signals:

![image](https://github.com/user-attachments/assets/365cbc09-fe6f-4f6b-833d-33831c6a3aea)

Unknown main signals:

![image](https://github.com/user-attachments/assets/4203f0bc-7ab8-401d-9ce0-482fc38dfb7d)

Unknown electricity signals:

![image](https://github.com/user-attachments/assets/8a910af6-9973-451b-bdf8-98c53d5a8fb0)