Let tile regions cover the entire planet (#287)
Part of https://github.com/hiddewie/OpenRailwayMap-vector/issues/231

This pull request expands the regions for which tiles are generated, to
cover the entire planet.

This means that any square-degree of planet that contains railways will
be served as tiles. Currently, only zoom 8+ is served if the tile is
within one of the existing regions.