Fix date missing from historical layers when loading map (#353)
Fixes https://github.com/hiddewie/OpenRailwayMap-vector/issues/350

Problem: during map load, the historical style is configured, before the
map is loaded or the date is changes. As such, there is no code that
configures the historical style to put the configured date into the
style layers.

Solution: if the historical view is active, configure the style to take
the date filtering into account.