Set replication timestamp in tile metadata (#358)
From
https://github.com/hiddewie/OpenRailwayMap-vector/issues/260#issuecomment-2713286861

The comment indicates that showing the freshness of the OSM data on
OpenRailwayMap is useful.

The replication timestamp is stored in the Osmium metadata during
download. of the OSM data. The Osmium metadata is imported into the
properties table of Osm2psql.

During tile generation the replication timestamp can be extracted and
stored in the tile metadata.

Merging and conversion to PMTiles should keep tile metadata intact.

Then, it has to be shown in the UI.