Fix POIs with zoom levels > 14 not output to tiles (#403)
Fixes #401 

The railway crossings, phones, subway entrances, buffer stops, derailer
and retarders were not output in tiles.

The tiles are generated up to and including zoom level 14. Higher zoom
levels are not output.

The standard POIs are now all output from zoom 14 and up. The map style
is conditioned on zoom levels, showing just the features that need to be
shown according to their minimum allowed zoom level.

Additionally, crossings are moved to zoom 15 instead of zoom 16.