Import station feature and state (#407)
This PR properly imports the feature (`railway` tag) and state of a
station.

For example a `disused:railway=yard` will be imported as `feature=yard`
and `state=disused`.

The UI can show the state, and render the station feature with a
different (gray?, brown?) color.

Razed station features will no longer be imported.