Replace reference to Americana style to Stadia Maps style (#433)
Fixes https://github.com/hiddewie/OpenRailwayMap-vector/discussions/349

The Americana style does not work in deployed domains. This project has
no affiliation with OSM US, so we drop the reference from the example in
the background styles. The example is replaced with Stadia Maps which
works cross domain, but requires a (free) API key.