Implement GB whistle boards (#523)
Adds [whistle boards](https://www.railsigns.uk/sect15page1.html) to
signals for GB.

`whistle-board.svg` and `whistle-stencil.svg` are compliant to the
standard save as to typeface (very close approximation).
`whistle-continuous.svg` is not drawn from a standard.

Photos of board and continuous can be found
[here](https://www.railsigns.uk/photos/p_whistle1.html#pic_wboard).

Note: I think that no `tags:` section is needed because otherwise it'd
just be duplicating tags unnecessarily in the data. If this is
technologically impossible please inform me and I will change as needed
to implement this.

---------

Co-authored-by: Hidde Wieringa <hidde@hiddewieringa.nl>