Add German signals Pl 3, Pl 4 (#564)
Added German railway signals Pl 3 and Pl 4 which are used on several
narrow-gauge railways in eastern Germany. Their documentation on the
wiki can be found
[here](https://wiki.openstreetmap.org/wiki/DE:OpenRailwayMap/Tagging_in_Germany/B%C3%BC,_Ne,_Pf_and_So_Signals#Pl-Signale),
images on Commons at [Pl
3](https://commons.wikimedia.org/wiki/Category:Pl_3_(German_railway_signal))/[Pl
4](https://commons.wikimedia.org/wiki/Category:Pl_4_(German_railway_signal)).

Examples for the signals on OSM are nodes
[13153203728](https://www.openstreetmap.org/node/13153203728) and
[9516827595](https://www.openstreetmap.org/node/9516827595),
respectively.

I built the SVGs from the existing symbols
[de/bue4-ds.svg](https://github.com/hiddewie/OpenRailwayMap-vector/blob/master/symbols/de/bue4-ds.svg)
and
[de/bue5.svg](https://github.com/hiddewie/OpenRailwayMap-vector/blob/master/symbols/de/bue5.svg)
with each part in their original size. Maybe this makes the combined
signal a bit too large?

This is my first pull request and I am not quite sure how it works; I
hope everything is okay.

---------

Co-authored-by: Hidde Wieringa <hidde@hiddewieringa.nl>