automatically generate taginfo.json file from the yaml files (#563)
Followup from
https://github.com/hiddewie/OpenRailwayMap-vector/pull/529#issuecomment-3293843431

This PR introduces a script, which combines the original `taginfo.json`
file with all the tags extracted from `features/*.yaml`. The combined
file is written to `./dist/taginfo.generated.json`, and the GitHub CI
'deploys' this to a branch called `dist`.

This means that we can manually define some tags in `taginfo.json`, and
the rest are automatically generated from the yaml files.

You can see an [example of the CI
logs](https://github.com/k-yle/OpenRailwayMap-vector/actions/runs/17912444557/job/50926863532),
and an [example of the generated
file](https://github.com/k-yle/OpenRailwayMap-vector/raw/dist/taginfo.generated.json).


The script is written in JavaScript since this project already has a lot
of JS, and it's the simplest option. But I could convert it to lua or
python if that's preferred.

---------

Co-authored-by: Hidde Wieringa <hidde@hiddewieringa.nl>